Module sentry::db

Re-exports

pub use self::campaign::*;

Modules

accounting

analytics

campaign

redis_pooltest-util

spendable

tests_postgrestest-util

validator_message

Structs

PostgresConfig

Connection configuration.

RedisError

Represents a redis error. For the most part you should be using the Error trait to interact with this rather than the actual struct.

TotalCount

Functions

get_channel_by_id

insert_channel

Used to insert/get Channel when creating a Campaign If channel already exists it will return it instead. This call should never trigger a SqlState::UNIQUE_VIOLATION

list_channels

Lists the Channels in ASC order.

postgres_connection

Uses the default max_size of the PoolConfig which is num_cpus::get_physical() * 4

redis_connection

setup_migrations

Sets the migrations using the POSTGRES_* environment variables

Type Definitions

DbPool

PoolError

Type alias for using [deadpool::managed::PoolError] with tokio_postgres.