Struct sentry::db::tests_postgres::Database

pub struct Database {
    pub name: String,
    pub pool: Pool,
}

Available on crate feature test-util only.

A Database is used to isolate test runs from each other we need to know the name of the database we’ve created. This will allow us the drop the database when we are recycling the connection

Fields
name: String

The database name that will be created by the pool CREATE DATABASE This database will be set on configuration level of the underlying connection Pool for tests

pool: Pool
Implementations

impl Database

pub fn new(name: String, pool: DbPool) -> Self