Function sentry::integrations::failure::panic_extractor

pub fn panic_extractor(info: &PanicInfo) -> Option<Event<'static>>

Extracts a Sentry Event from a PanicInfo.

Creates a new Sentry Event when the panic has a failure::Error payload. This is for use with the sentry-panic integration, and is enabled by default in sentry.

Examples

let panic_integration =
    sentry_panic::PanicIntegration::new().add_extractor(sentry_failure::panic_extractor);