Trait sentry::integrations::failure::FailureResultExt

pub trait FailureResultExt {
    type Value;
    fn fallible_unwrap(self) -> Self::Value;
}

Extension trait providing methods to unwrap a result, preserving backtraces from the underlying error in the event of a panic.

Associated Types

type Value

Type of the success case

Loading content...
Required methods

fn fallible_unwrap(self) -> Self::Value

Unwraps the result, panicking if it contains an error. Any backtrace attached to the error will be preserved with the panic.

Loading content...