Struct sentry::integrations::log::LoggerOptions

pub struct LoggerOptions {
    pub global_filter: Option<LevelFilter>,
    pub filter: LevelFilter,
    pub emit_breadcrumbs: bool,
    pub emit_error_events: bool,
    pub emit_warning_events: bool,
}

Logger specific options.

Fields
global_filter: Option<LevelFilter>

The global filter that should be used (also used before dispatching to the nested logger).

filter: LevelFilter

The sentry specific log level filter (defaults to Info)

emit_breadcrumbs: bool

If set to true, breadcrumbs are emitted. (defaults to true)

emit_error_events: bool

If set to true error events are sent for errors in the log. (defaults to true)

emit_warning_events: bool

If set to true warning events are sent for warnings in the log. (defaults to false)