Module sentry::internals

Useful internals.

This module contains types that users of the create typically do not have to directly interface with directly. These are often returned from methods on other types.

Structs

Auth
Represents an auth header.

ChronoParseError
An error from the parse function.

ClientInitGuard
Helper struct that is returned from init.

DateTime
ISO 8601 combined date and time with time zone.

DebugId
Unique identifier for debug information files and their debug information.

Dsn
Represents a Sentry dsn.

ParseDebugIdError
Indicates a parsing error

ProjectId
Represents a project ID.

ScopeGuard
A scope guard.

Utc
The UTC time zone. This is the most efficient time zone when you don't need the local time. It is also used as an offset (which is also a dummy type).

Uuid
A Universally Unique Identifier (UUID).

Enums

DsnParseError
Represents a dsn url parsing error.

ProjectIdParseError
Raised if a project ID cannot be parsed from a string.

Scheme
Represents the scheme of an url http/https.

UuidVariant
The reserved variants of UUIDs.

UuidVersion
The version of the UUID, denoting the generating algorithm.

Traits

IntoBreadcrumbs
A helper trait that converts an object into a breadcrumb.

IntoDsn
Helper trait to convert a string into an Option<Dsn>.

TimeZone
The time zone.

Transport
The trait for transports.

TransportFactory
A factory creating transport instances.