Struct sentry::internals::ProjectId

pub struct ProjectId { /* fields omitted */ }

Represents a project ID.

This is a thin wrapper around IDs supported by the Sentry server. The idea is that the sentry server generally can switch the ID format in the future (eg: we implement the IDs as strings and not as integers) but the actual ID format that is encountered are currently indeed integers.

To be future proof we support either integers or "short" strings.