Trait sentry::internals::IntoDsn

pub trait IntoDsn {
    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError>;
}

Helper trait to convert a string into an Option<Dsn>.

This converts a value into a DSN by parsing. The empty string or null values result in no DSN being parsed.

Required methods

fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError>

Converts the value into a Result<Option<Dsn>, E>.

Loading content...