Trait sentry::internals::Transport

pub trait Transport: 'static + Send + Sync {
    fn send_event(&self, event: Event<'static>);

    fn shutdown(&self, timeout: Duration) -> bool { ... }
}

The trait for transports.

A transport is responsible for sending events to Sentry. Custom implementations can be created to use a different abstraction to send events. This is for instance used for the test system.

Required methods

fn send_event(&self, event: Event<'static>)

Sends an event.

Loading content...
Provided methods

fn shutdown(&self, timeout: Duration) -> bool

Drains the queue if there is one.

The default implementation does nothing. If the queue was successfully shutdowned the return value should be true or false if events were left in it.

Loading content...