Module sentry::protocol

The current latest sentry protocol version.

Most constructs in the protocol map directly to types here but some cleanup by renaming attributes has been applied. The idea here is that a future sentry protocol will be a cleanup of the old one and is mapped to similar values on the rust side.

Modules

debugid
Represents a debug ID.

map
The internally used arbitrary data map type.

value
An arbitrary (JSON) value.

Structs

Addr
Represents an address.

AppContext
Holds app information.

AppleDebugImage
Represents an apple debug image in the debug meta.

Breadcrumb
Represents a single breadcrumb.

BrowserContext
Holds information about the web browser.

CError
POSIX signal with optional extended data.

ClientSdkInfo
Information on the SDK client.

ClientSdkPackage
Represents an installed package relevant to the SDK.

DebugMeta
Represents debug meta information.

DeviceContext
Holds device information.

Event
Represents a full event for Sentry.

Exception
Represents a single exception.

Frame
Represents a frame.

LogEntry
Represents a log entry message.

MachException
Mach exception information.

Map
A map based on a B-Tree.

Mechanism
Represents a single exception.

MechanismMeta
Operating system or runtime meta information to an exception mechanism.

OsContext
Holds operating system information.

ParseLevelError
An error used when parsing Level.

PosixSignal
POSIX signal with optional extended data.

ProguardDebugImage
Represents a proguard mapping file reference.

RegVal
Represents a register value.

Request
Represents http request data.

RuntimeContext
Holds information about the runtime.

Stacktrace
Represents a stacktrace.

SymbolicDebugImage
Represents a symbolic debug image.

SystemSdkInfo
Holds information about the system SDK.

TemplateInfo
Represents template debug info.

Thread
Represents a single thread.

User
Represents user info.

Values
A wrapper type for collections with attached meta data.

Enums

Context
Typed contextual data.

DebugImage
Represents a debug image.

IpAddress
An IP address, either IPv4, IPv6 or Auto.

Level
Represents the level of severity of an event or breadcrumb.

Orientation
Optional device screen orientation

ThreadId
Represents a thread id.

Value
Represents any valid JSON value.