Module sentry::protocol::value

An arbitrary (JSON) value.

Structs

Map
Represents a JSON key/value type.

Number
Represents a JSON number, whether integer or floating point.

Enums

Value
Represents any valid JSON value.

Traits

Index
A type that can be used to index into a serde_json::Value.

Functions

from_value
Interpret a serde_json::Value as an instance of type T.

to_value
Convert a T into serde_json::Value which is an enum that can represent any valid JSON data.