Module sentry::routes

Sentry REST API documentation

All routes are listed below. Here is an overview and links to all of them:

Channel routes

GET /v5/channel/list

GET /v5/channel/:id/accounting

GET /v5/channel/:id/spender/:addr (auth required)

POST /v5/channel/:id/spender/:addr (auth required)

GET /v5/channel/:id/spender/all (auth required)

GET /v5/channel/:id/validator-messages

GET /v5/channel/:id/validator-messages/:addr

GET /v5/channel/:id/validator-messages/:addr/:validator_messages

POST /v5/channel/:id/validator-messages (auth required)

GET /v5/channel/:id/last-approved

POST /v5/channel/:id/pay (auth required)

GET /v5/channel/:id/get-leaf

POST /v5/channel/dummy-deposit (auth required) available only with Dummy adapter

Campaign routes

GET /v5/campaign/list

POST /v5/campaign (auth required)

POST /v5/campaign/:id (auth required)

POST /v5/campaign/:id/events (auth required)

POST /v5/campaign/:id/close (auth required)

Analytics routes

GET /v5/analytics

GET /v5/analytics/for-publisher (auth required)

GET /v5/analytics/for-advertiser (auth required)

GET /v5/analytics/for-admin (auth required)

Channel

All routes are implemented under the module channel.

Route parameters

Paths which include these parameters are validated as follows:

:id - ChannelId

:addr - a valid Address or ValidatorId.

Routes

GET /v5/channel/list

The route is handled by channel::channel_list().

Request query parameters: ChannelListQuery

Response: ChannelListResponse

Examples

Query:

use primitives::{
    sentry::channel_list::ChannelListQuery,
    test_util::{IDS, LEADER},
    ChainId,
};

fn main() {
    // An empty query
    {
        let empty = "";
        let empty_expected = ChannelListQuery {
            page: 0,
            validator: None,
            chains: vec![],
        };

        assert_eq!(empty_expected, serde_qs::from_str(empty).unwrap());
    }

    // Query with `page`
    {
        let only_page = "page=14";
        let only_page_expected = ChannelListQuery {
            page: 14,
            validator: None,
            chains: vec![],
        };

        assert_eq!(only_page_expected, serde_qs::from_str(only_page).unwrap());
    }

    // Query with `validator`
    {
        let only_validator = "validator=0x80690751969B234697e9059e04ed72195c3507fa";
        let only_validator_expected = ChannelListQuery {
            page: 0,
            validator: Some(IDS[&LEADER]),
            chains: vec![],
        };

        assert_eq!(
            only_validator_expected,
            serde_qs::from_str(only_validator).unwrap()
        );
    }

    // Query with `chains`
    {
        let chains_query = "chains[]=1&chains[]=1337";
        let chains_expected = ChannelListQuery {
            page: 0,
            validator: None,
            chains: vec![ChainId::new(1), ChainId::new(1337)],
        };

        assert_eq!(chains_expected, serde_qs::from_str(chains_query).unwrap());
    }

    // Query with all parameters
    {
        let all_query =
            "page=14&validator=0x80690751969B234697e9059e04ed72195c3507fa&chains[]=1&chains[]=1337";
        let all_expected = ChannelListQuery {
            page: 14,
            validator: Some(IDS[&LEADER]),
            chains: vec![ChainId::new(1), ChainId::new(1337)],
        };

        assert_eq!(all_expected, serde_qs::from_str(all_query).unwrap());
    }
}

GET /v5/channel/:id/accounting

The route is handled by channel::get_accounting_for_channel().

Response: AccountingResponse::<CheckedState>

GET /v5/channel/:id/spender/:addr (auth required)

The route is handled by channel::get_spender_limits().

Response: SpenderResponse

POST /v5/channel/:id/spender/:addr (auth required)

This route forces the addition of a spender Accounting (if one does not exist) to the given Channel with spent = 0. This will also ensure that the spender is added to the NewState as well.

The route is handled by channel::add_spender_leaf().

Response: SuccessResponse

GET /v5/channel/:id/spender/all (auth required)

The route is handled by channel::get_all_spender_limits().

Response: AllSpendersResponse

GET /v5/channel/:id/validator-messages

GET /v5/channel/:id/validator-messages/:addr - filter by the given ValidatorId

GET /v5/channel/:id/validator-messages/:addr/:validator_messages - filters by the given ValidatorId and a Validator message types.

:validator_messages - url encoded list of Validator message types separated by a +.

E.g. NewState+ApproveState becomes NewState%2BApproveState

The route is handled by channel::validator_message::list_validator_messages().

Request query parameters: ValidatorMessagesListQuery

Response: ValidatorMessagesListResponse

POST /v5/channel/:id/validator-messages (auth required)

The route is handled by channel::validator_message::create_validator_messages().

Request body (json): ValidatorMessagesCreateRequest

Example:

{
    "messages": [
        /// validator messages
        ...
    ]
}

Validator messages: MessageTypes

GET /v5/channel/:id/last-approved

The route is handled by channel::last_approved().

Request query parameters: LastApprovedQuery

Response: LastApprovedResponse

POST /v5/channel/:id/pay (auth required)

Channel Payout with authentication of the spender.

This route handles withdrawals of advertiser funds for the authenticated spender. It needs to ensure all campaigns are closed. It accepts a JSON body in the request which contains all of the earners and updates their balances accordingly. Used when an advertiser/spender wants to get their remaining funds back.

The route is handled by channel::channel_payout().

Request JSON body: ChannelPayRequest

Response: SuccessResponse

GET /v5/channel/:id/get-leaf

TODO: implement and document as part of issue #382

This route gets the latest approved state (NewState/ApproveState pair), and finds the given spender/earner in the balances tree, and produce a merkle proof for it. This is useful for the Platform to verify if a spender leaf really exists.

The route is handled by todo.

Request query parameters:

spender=[0x...] or earner=[0x...] (required)

Example Spender:

/get-leaf?spender=0x...

Example Earner:

/get-leaf?earner=0x.... This module includes all routes for Sentry and the documentation of each Request/Response.

POST /v5/channel/dummy-deposit (auth required)

Set a deposit for a Channel and depositor (the authenticated address) in the Dummy adapter.

NOTE: This route is available only when using the Dummy adapter.

The route is handled by channel::channel_dummy_deposit().

Request body (json): ChannelDummyDeposit

Response: SuccessResponse

Campaign

All routes are implemented under the module campaign.

Route parameters

Paths which include these parameters are validated as follow:

:id - CampaignId

Routes

GET /v5/campaign/list

Lists all campaigns with pagination and orders them in ascending order (ASC) by Campaign.created. This ensures that the order in the pages will not change if a new Campaign is created while still retrieving a page.

The route is handled by campaign::campaign_list().

Request query parameters: CampaignListQuery

page=[integer] (optional) default: 0

creator=[0x....] (optional) - address of the creator to be filtered by

activeTo=[integer] (optional) in seconds - filters campaigns by Campaign.active.to > query.activeTo

validator=[0x...] or leader=[0x...] (optional) - address of the validator to be filtered by. You can either

validator=[0x...] - it will return all Campaigns where this address is either Channel.leader or Channel.follower

leader=[0x...] - it will return all Campaigns where this address is Channel.leader

Response: CampaignListResponse

Examples

Query:

use chrono::{TimeZone, Utc};
use primitives::{
    sentry::campaign_list::{CampaignListQuery, ValidatorParam},
    test_util::{ADVERTISER, FOLLOWER, IDS, LEADER},
};

fn main() {
    // Empty query - default values only
    {
        let empty_query = "";
        let query: CampaignListQuery = serde_qs::from_str(empty_query).unwrap();

        assert_eq!(0, query.page);
        assert!(
            Utc::now() >= query.active_to_ge,
            "By default `activeTo` is set to `Utc::now()`"
        );
        assert!(query.creator.is_none());
        assert!(query.validator.is_none());
    }

    // In the following examples we always use `activeTo`
    // as it makes simpler examples for assertions rather than using the default `Utc::now()`

    // Query with `activeTo` only
    {
        let active_to_query = "activeTo=1624192200";
        let active_to = CampaignListQuery {
            page: 0,
            active_to_ge: Utc.ymd(2021, 6, 20).and_hms(12, 30, 0),
            creator: None,
            validator: None,
        };

        assert_eq!(active_to, serde_qs::from_str(active_to_query).unwrap());
    }

    // Query with `page` & `activeTo`
    {
        let with_page_query = "page=14&activeTo=1624192200";
        let with_page = CampaignListQuery {
            page: 14,
            active_to_ge: Utc.ymd(2021, 6, 20).and_hms(12, 30, 0),
            creator: None,
            validator: None,
        };

        assert_eq!(with_page, serde_qs::from_str(with_page_query).unwrap());
    }

    // Query with `creator`
    {
        let with_creator_query =
            "activeTo=1624192200&creator=0xDd589B43793934EF6Ad266067A0d1D4896b0dff0";

        let with_creator = CampaignListQuery {
            page: 0,
            active_to_ge: Utc.ymd(2021, 6, 20).and_hms(12, 30, 0),
            creator: Some(*ADVERTISER),
            validator: None,
        };

        assert_eq!(
            with_creator,
            serde_qs::from_str(with_creator_query).unwrap()
        );
    }

    // Query with `validator`
    // You can either have `leader` or `validator` but not both!
    {
        let with_creator_validator_query =
            "activeTo=1624192200&validator=0xf3f583AEC5f7C030722Fe992A5688557e1B86ef7";
        let with_creator_validator = CampaignListQuery {
            page: 0,
            active_to_ge: Utc.ymd(2021, 6, 20).and_hms(12, 30, 0),
            creator: None,
            validator: Some(ValidatorParam::Validator(IDS[&FOLLOWER])),
        };

        assert_eq!(
            with_creator_validator,
            serde_qs::from_str(with_creator_validator_query).unwrap()
        );
    }

    // Query with `leader`
    // You can either have `leader` or `validator` but not both!
    {
        let with_leader_query =
            "activeTo=1624192200&leader=0x80690751969B234697e9059e04ed72195c3507fa";

        let with_leader = CampaignListQuery {
            page: 0,
            active_to_ge: Utc.ymd(2021, 6, 20).and_hms(12, 30, 0),
            creator: None,
            validator: Some(ValidatorParam::Leader(IDS[&LEADER])),
        };

        assert_eq!(with_leader, serde_qs::from_str(with_leader_query).unwrap());
    }

    // Query with all parameters and `validator`
    // You can either have `leader` or `validator` but not both!
    {
        let full_query = "page=14&activeTo=1624192200&creator=0xDd589B43793934EF6Ad266067A0d1D4896b0dff0&validator=0xf3f583AEC5f7C030722Fe992A5688557e1B86ef7";
        let full_expected = CampaignListQuery {
            page: 14,
            active_to_ge: Utc.ymd(2021, 6, 20).and_hms(12, 30, 0),
            creator: Some(*ADVERTISER),
            validator: Some(ValidatorParam::Validator(IDS[&FOLLOWER])),
        };

        assert_eq!(full_expected, serde_qs::from_str(full_query).unwrap());
    }
}

Response:

use primitives::sentry::campaign_list::CampaignListResponse;
use serde_json::{from_value, json};

fn main() {
    let json = json!({
      "campaigns": [
        {
          "id": "0x936da01f9abd4d9d80c702af85c822a8",
          "channel": {
            "leader": "0x80690751969B234697e9059e04ed72195c3507fa",
            "follower": "0xf3f583AEC5f7C030722Fe992A5688557e1B86ef7",
            "guardian": "0xe061E1EB461EaBE512759aa18A201B20Fe90631D",
            "token": "0x2BCaf6968aEC8A3b5126FBfAb5Fd419da6E8AD8E",
            "nonce": "0"
          },
          "creator": "0xDd589B43793934EF6Ad266067A0d1D4896b0dff0",
          "budget": "15000000000",
          "validators": [
            {
              "id": "0x80690751969B234697e9059e04ed72195c3507fa",
              "fee": "500000000",
              "url": "http://localhost:8005/"
            },
            {
              "id": "0xf3f583AEC5f7C030722Fe992A5688557e1B86ef7",
              "fee": "400000000",
              "url": "http://localhost:8006/"
            }
          ],
          "title": "Dummy Campaign",
          "pricingBounds": {
            "CLICK": {
              "min": "6000000",
              "max": "10000000"
            },
            "IMPRESSION": {
              "min": "4000000",
              "max": "5000000"
            }
          },
          "eventSubmission": {
            "allow": []
          },
          "adUnits": [
            {
              "ipfs": "Qmasg8FrbuSQpjFu3kRnZF9beg8rEBFrqgi1uXDRwCbX5f",
              "type": "legacy_250x250",
              "mediaUrl": "ipfs://QmcUVX7fvoLMM93uN2bD3wGTH8MXSxeL8hojYfL2Lhp7mR",
              "mediaMime": "image/jpeg",
              "targetUrl": "https://www.adex.network/?stremio-test-banner-1",
              "owner": "0xE882ebF439207a70dDcCb39E13CA8506c9F45fD9",
              "created": 1564390800000_u64,
              "title": "Dummy AdUnit 1",
              "description": "Dummy AdUnit description 1",
              "archived": false
            },
            {
              "ipfs": "QmVhRDGXoM3Fg3HZD5xwMuxtb9ZErwC8wHt8CjsfxaiUbZ",
              "type": "legacy_250x250",
              "mediaUrl": "ipfs://QmQB7uz7Gxfy7wqAnrnBcZFaVJLos8J9gn8mRcHQU6dAi1",
              "mediaMime": "image/jpeg",
              "targetUrl": "https://www.adex.network/?adex-campaign=true&pub=stremio",
              "owner": "0xE882ebF439207a70dDcCb39E13CA8506c9F45fD9",
              "created": 1564390800000_u64,
              "title": "Dummy AdUnit 2",
              "description": "Dummy AdUnit description 2",
              "archived": false
            }
          ],
          "targetingRules": [],
          "created": 1612162800000_u64,
          "activeTo": 4073414400000_u64
        },
        {
          "id": "0x127b98248f4e4b73af409d10f62daeaa",
          "channel": {
            "leader": "0xf3f583AEC5f7C030722Fe992A5688557e1B86ef7",
            "follower": "0x80690751969B234697e9059e04ed72195c3507fa",
            "guardian": "0x79D358a3194d737880B3eFD94ADccD246af9F535",
            "token": "0x2BCaf6968aEC8A3b5126FBfAb5Fd419da6E8AD8E",
            "nonce": "0"
          },
          "creator": "0xDd589B43793934EF6Ad266067A0d1D4896b0dff0",
          "budget": "2000000000",
          "validators": [
            {
              "id": "0xf3f583AEC5f7C030722Fe992A5688557e1B86ef7",
              "fee": "10000000",
              "url": "http://localhost:8006/"
            },
            {
              "id": "0x80690751969B234697e9059e04ed72195c3507fa",
              "fee": "5000000",
              "url": "http://localhost:8005/"
            }
          ],
          "title": "Dummy Campaign 2 in Chain #1337",
          "pricingBounds": {
            "CLICK": {
              "min": "300000000",
              "max": "500000000"
            },
            "IMPRESSION": {
              "min": "100000000",
              "max": "200000000"
            }
          },
          "eventSubmission": {
            "allow": []
          },
          "adUnits": [
            {
              "ipfs": "Qmasg8FrbuSQpjFu3kRnZF9beg8rEBFrqgi1uXDRwCbX5f",
              "type": "legacy_250x250",
              "mediaUrl": "ipfs://QmcUVX7fvoLMM93uN2bD3wGTH8MXSxeL8hojYfL2Lhp7mR",
              "mediaMime": "image/jpeg",
              "targetUrl": "https://www.adex.network/?stremio-test-banner-1",
              "owner": "0xE882ebF439207a70dDcCb39E13CA8506c9F45fD9",
              "created": 1564390800000_u64,
              "title": "Dummy AdUnit 1",
              "description": "Dummy AdUnit description 1",
              "archived": false
            },
            {
              "ipfs": "QmVhRDGXoM3Fg3HZD5xwMuxtb9ZErwC8wHt8CjsfxaiUbZ",
              "type": "legacy_250x250",
              "mediaUrl": "ipfs://QmQB7uz7Gxfy7wqAnrnBcZFaVJLos8J9gn8mRcHQU6dAi1",
              "mediaMime": "image/jpeg",
              "targetUrl": "https://www.adex.network/?adex-campaign=true&pub=stremio",
              "owner": "0xE882ebF439207a70dDcCb39E13CA8506c9F45fD9",
              "created": 1564390800000_u64,
              "title": "Dummy AdUnit 2",
              "description": "Dummy AdUnit description 2",
              "archived": false
            }
          ],
          "targetingRules": [],
          "created": 1612162800000_u64,
          "activeFrom": 1656280800000_u64,
          "activeTo": 4073414400000_u64
        },
        {
          "id": "0xa78f3492481b41a688488a7aa1ff17df",
          "channel": {
            "leader": "0x80690751969B234697e9059e04ed72195c3507fa",
            "follower": "0xf3f583AEC5f7C030722Fe992A5688557e1B86ef7",
            "guardian": "0x79D358a3194d737880B3eFD94ADccD246af9F535",
            "token": "0x12a28f2bfBFfDf5842657235cC058242f40fDEa6",
            "nonce": "1"
          },
          "creator": "0x541b401362Ea1D489D322579552B099e801F3632",
          "budget": "2000000000",
          "validators": [
            {
              "id": "0x80690751969B234697e9059e04ed72195c3507fa",
              "fee": "200000000",
              "url": "http://localhost:8005/"
            },
            {
              "id": "0xf3f583AEC5f7C030722Fe992A5688557e1B86ef7",
              "fee": "175000000",
              "url": "http://localhost:8006/"
            }
          ],
          "title": "Dummy Campaign 3 in Chain #1",
          "pricingBounds": {
            "CLICK": {
              "min": "3500000",
              "max": "6500000"
            },
            "IMPRESSION": {
              "min": "1500000",
              "max": "2500000"
            }
          },
          "eventSubmission": {
            "allow": []
          },
          "adUnits": [
            {
              "ipfs": "QmYwcpMjmqJfo9ot1jGe9rfXsszFV1WbEA59QS7dEVHfJi",
              "type": "legacy_250x250",
              "mediaUrl": "ipfs://QmQB7uz7Gxfy7wqAnrnBcZFaVJLos8J9gn8mRcHQU6dAi1",
              "mediaMime": "image/jpeg",
              "targetUrl": "https://www.adex.network/?adex-campaign=true",
              "owner": "0xE882ebF439207a70dDcCb39E13CA8506c9F45fD9",
              "created": 1564390800000_u64,
              "title": "Dummy AdUnit 3",
              "description": "Dummy AdUnit description 3",
              "archived": false
            },
            {
              "ipfs": "QmTAF3FsFDS7Ru8WChoD9ofiHTH8gAQfR4mYSnwxqTDpJH",
              "type": "legacy_250x250",
              "mediaUrl": "ipfs://QmQAcfBJpDDuH99A4p3pFtUmQwamS8UYStP5HxHC7bgYXY",
              "mediaMime": "image/jpeg",
              "targetUrl": "https://adex.network",
              "owner": "0xE882ebF439207a70dDcCb39E13CA8506c9F45fD9",
              "created": 1564390800000_u64,
              "title": "Dummy AdUnit 4",
              "description": "Dummy AdUnit description 4",
              "archived": false
            }
          ],
          "targetingRules": [],
          "created": 1612162800000_u64,
          "activeTo": 4073414400000_u64
        }
      ],
      "totalPages": 1,
      "page": 0
    });

    assert!(from_value::<CampaignListResponse>(json).is_ok());
}

POST /v5/campaign (auth required)

Create a new Campaign. Request must be sent by the Campaign.creator.

Authentication is required to validate Campaign.creator == Auth.uid

It will make sure the Channel is created if new and it will update the spendable amount using the Adapter::get_deposit().

The route is handled by campaign::create_campaign().

Request body (json): CreateCampaign

Response: Campaign

POST /v5/campaign/:id (auth required)

Modify the Campaign. Request must be sent by the Campaign.creator.

Authentication is required to validate Campaign.creator == Auth.uid

The route is handled by campaign::update_campaign::handle_route().

Request body (json): ModifyCampaign

Response: Campaign

POST /v5/campaign/:id/events

Add new Events (IMPRESSIONs & CLICKs) to the Campaign. Applies Campaign.event_submission rules and additional validation using check_access().

The route is handled by campaign::insert_events::handle_route().

Request body (json):

{
    "events": [
        // Events
   ]
}

Response: SuccessResponse

POST /v5/campaign/:id/close (auth required)

Close the campaign.

The route is handled by campaign::close_campaign().

Request must be sent by the Campaign.creator.

Authentication is required to validate Campaign.creator == Auth.uid

Closes the campaign by setting Campaign.budget so that remaining budget = 0.

Response: SuccessResponse

Analytics

GET /v5/analytics

Allowed keys: AllowedKey::Country, AllowedKey::AdSlotType

GET /v5/analytics/for-publisher (auth required)

Returns all analytics where the currently authenticated address Auth.uid is a publisher.

All ALLOWED_KEYS are allowed for this route.

The route is handled by get_analytics().

GET /v5/analytics/for-advertiser (auth required)

Returns all analytics where the currently authenticated address Auth.uid is an advertiser.

All ALLOWED_KEYS are allowed for this route.

The route is handled by get_analytics().

GET /v5/analytics/for-admin (auth required)

Admin access to the analytics with no restrictions on the keys for filtering.

All ALLOWED_KEYS are allowed for admins.

Admin addresses are configured in the Config.admins.

The route is handled by get_analytics().

Modules

campaign

/v5/campaign routes

channel

/v5/channel routes

routers

This module contains all the Sentry REST API routers.

Functions

get_analytics

GET /v5/analytics request with query parameters: primitives::analytics::AnalyticsQuery.

get_cfg

GET /cfg request