Function sentry::routes::campaign::update_campaign::handle_route

pub async fn handle_route<C: Locked + 'static>(
    __arg0: Json<ModifyCampaign>,
    __arg1: Extension<ChainOf<Campaign>>,
    __arg2: Extension<Arc<Application<C>>>
) -> Result<Json<Campaign>, ResponseError>

POST /v5/campaign/:id (auth required)

Request body (json): ModifyCampaign consists of all of the editable fields of the [Campaign]

ResponseCampaign

Ensures that the remaining funds for all campaigns <= total remaining funds (total deposited - total spent)