Function sentry::test::with_captured_events_options

pub fn with_captured_events_options<F, O>(
    f: F, 
    options: O
) -> Vec<Event<'static>> where
    F: FnOnce(),
    O: Into<ClientOptions>, 

Runs some code with the default test hub with the given options and returns the captured events.

If not DSN is set on the options a default test DSN is inserted. The transport on the options is also overridden with a TestTransport.

This is a shortcut for creating a testable hub with the supplied options and to call run_and_capture_events on it.