Trait sentry::SentryFutureExt

pub trait SentryFutureExt {
    fn bind_hub<H>(self, hub: H) -> SentryFuture<Self>
    where
        H: Into<Arc<Hub>>,
    { ... }
}

Future extensions for Sentry.

Provided Methods

fn bind_hub<H>(self, hub: H) -> SentryFuture<Self> where
H: Into<Arc<Hub>>,

Binds a hub to the execution of this future.

This ensures that the future is polled within the given hub.