Function sentry::integrations::failure::exception_from_single_fail

pub fn exception_from_single_fail<F>(f: &F, bt: Option<&Backtrace>) -> Exception where
    F: Fail + ?Sized, 

This converts a single fail instance into an exception.

This is typically not very useful as the event_from_error and event_from_fail methods will assemble an entire event with all the causes of a failure, however for certain more complex situations where fails are contained within a non fail error type that might also carry useful information it can be useful to call this method instead.