Struct sentry::integrations::log::LoggerOptions

pub struct LoggerOptions {
    pub global_filter: Option<LevelFilter>,
    pub filter: LevelFilter,
    pub emit_breadcrumbs: bool,
    pub emit_error_events: bool,
    pub emit_warning_events: bool,
    pub attach_stacktraces: bool,
}

Logger specific options.

Fields
global_filter: Option<LevelFilter>

The global filter that should be used (also used before dispatching to the nested logger).

filter: LevelFilter

The sentry specific log level filter (defaults to Info)

emit_breadcrumbs: bool

If set to true, breadcrumbs will be emitted. (defaults to true)

emit_error_events: bool

If set to true error events will be sent for errors in the log. (defaults to true)

emit_warning_events: bool

If set to true warning events will be sent for warnings in the log. (defaults to false)

attach_stacktraces: bool

If set to true current stacktrace will be resolved and attached to each event. (expensive, defaults to true)