Module sentry::integrations::panic

Panic handler support.

Feature: with_panic (enabled by default)

A panic handler can be installed that will automatically dispatch all errors to Sentry that are caused by a panic.

Configuration

use sentry::integrations::panic::register_panic_handler;
register_panic_handler();

Additionally panics are forwarded to the previously registered panic hook.

Functions

event_from_panic_info
Creates an event from the given panic info.

message_from_panic_info
Extract the message of a panic.

panic_handler
A panic handler that sends to Sentry.

register_panic_handler
Registes the panic handler.