Module sentry::internals

👎 Deprecated:
These exports have been moved to the root or the types mod.

Useful internals.

This module contains types that users of the crate typically do not have to interface with directly. These are often returned from methods on other types.

Modules

protocol

This module exposes the types for the Sentry protocol in different versions.

Structs

Auth

Represents an auth header.

BreakpadFormat

Wrapper around DebugId for Breakpad formatting.

ChronoParseError

An error from the parse function.

ClientInitGuard

Helper struct that is returned from init.

CodeId

Unique platform-dependent identifier of code files.

DateTime

ISO 8601 combined date and time with time zone.

DebugId

Unique identifier for debug information files and their debug information.

Dsn

Represents a Sentry dsn.

ParseCodeIdError

Indicates an error parsing a CodeId.

ParseDebugIdError

Indicates an error parsing a DebugId.

ProjectId

Represents a project ID.

Utc

The UTC time zone. This is the most efficient time zone when you don’t need the local time. It is also used as an offset (which is also a dummy type).

Uuid

A Universally Unique Identifier (UUID).

Enums

ParseAuthError

Represents an auth header parsing error.

ParseDsnError

Represents a dsn url parsing error.

ParseProjectIdError

Raised if a project ID cannot be parsed from a string.

Scheme

Represents the scheme of an url http/https.

UuidVariant

The reserved variants of UUIDs.

UuidVersion

The version of the UUID, denoting the generating algorithm.

Traits

IntoBreadcrumbs

A helper trait that converts self into an Iterator of Breadcrumbs.

IntoDsn

Helper trait to convert a string into an Option<Dsn>.

TimeZone

The time zone.

Transport

The trait for transports.

TransportFactory

A factory creating transport instances.

Functions

apply_defaults

Apply default client options.