Trait sentry::internals::IntoDsn

pub trait IntoDsn {
    fn into_dsn(self) -> Result<Option<Dsn>, DsnParseError>;
}

Helper trait to convert a string into an Option<Dsn>.

This converts a value into a DSN by parsing. The empty string or null values result in no DSN being parsed.

Required Methods

fn into_dsn(self) -> Result<Option<Dsn>, DsnParseError>

Converts the value into a Result<Option<Dsn>, E>.