Module sentry::routes

Sentry REST API documentation

Channel

All routes are implemented under the module channel.

Route parameters

Paths which include these parameters are validated as follows:

:id - ChannelId

:addr - a valid Address or ValidatorId.

Routes

GET /v5/channel/list

Request query parameters: ChannelListQuery

Response: ChannelListResponse

GET /v5/channel/:id/accounting

Response: AccountingResponse::<CheckedState>

GET /v5/channel/:id/spender/:addr (auth required)

Response: SpenderResponse

POST /v5/channel/:id/spender/:addr (auth required)

This route forces the addition of a spender Accounting (if one does not exist) to the given Channel with spent = 0. This will also ensure that the spender is added to the NewState as well.

Response: SuccessResponse

GET /v5/channel/:id/spender/all (auth required)

Response: AllSpendersResponse

GET /v5/channel/:id/validator-messages

GET /v5/channel/:id/validator-messages/:addr - filter by the given ValidatorId

GET /v5/channel/:id/validator-messages/:addr/:validator_messages - filters by the given ValidatorId and a Validator message types.

:validator_messages - url encoded list of Validator message types separated by a +.

E.g. NewState+ApproveState becomes NewState%2BApproveState

Request query parameters: ValidatorMessagesListQuery

Response: ValidatorMessagesListResponse

POST /v5/channel/:id/validator-messages (auth required)

Request body (json): ValidatorMessagesCreateRequest

Example:

{
    "messages": [
        /// validator messages
        ...
    ]
}

Validator messages: MessageTypes

GET /v5/channel/:id/last-approved

Request query parameters: LastApprovedQuery

Response: LastApprovedResponse

POST /v5/channel/:id/pay (auth required)

Channel Payout with authentication of the spender.

This route handles withdrawals of advertiser funds for the authenticated spender. It needs to ensure all campaigns are closed. It accepts a JSON body in the request which contains all of the earners and updates their balances accordingly. Used when an advertiser/spender wants to get their remaining funds back.

Request JSON body: ChannelPayRequest

Response: SuccessResponse

GET /v5/channel/:id/get-leaf

TODO: implement and document as part of issue #382

This route gets the latest approved state (NewState/ApproveState pair), and finds the given spender/earner in the balances tree, and produce a merkle proof for it. This is useful for the Platform to verify if a spender leaf really exists.

Request query parameters:

spender=[0x...] or earner=[0x...] (required)

Example Spender:

/get-leaf?spender=0x...

Example Earner:

/get-leaf?earner=0x.... This module includes all routes for Sentry and the documentation of each Request/Response.

Campaign

All routes are implemented under the module campaign.

Route parameters

Paths which include these parameters are validated as follow:

:id - CampaignId

Routes

GET /v5/campaign/list

Lists all campaigns with pagination and orders them in descending order (DESC) by Campaign.created. This ensures that the order in the pages will not change if a new Campaign is created while still retrieving a page.

Request query parameters: CampaignListQuery

page=[integer] (optional) default: 0

creator=[0x....] (optional) - address of the creator to be filtered by

activeTo=[integer] (optional) in seconds - filters campaigns by Campaign.active.to > query.activeTo

validator=[0x...] or leader=[0x...] (optional) - address of the validator to be filtered by. You can either

validator=[0x...] - it will return all Campaigns where this address is either Channel.leader or Channel.follower

leader=[0x...] - it will return all Campaigns where this address is Channel.leader

Response: CampaignListResponse

POST /v5/campaign (auth required)

Create a new Campaign. Request must be sent by the Campaign.creator.

Authentication is required to validate Campaign.creator == Auth.uid

It will make sure the Channel is created if new and it will update the spendable amount using the Adapter::get_deposit().

Request body (json): CreateCampaign

Response: Campaign

POST /v5/campaign/:id (auth required)

Modify the Campaign. Request must be sent by the Campaign.creator.

Authentication is required to validate Campaign.creator == Auth.uid

Request body (json): ModifyCampaign

Response: Campaign

POST /v5/campaign/:id/close (auth required)

Close the campaign.

Request must be sent by the Campaign.creator.

Authentication is required to validate Campaign.creator == Auth.uid

Closes the campaign by setting Campaign.budget so that remaining budget = 0.

Response: SuccessResponse

Analytics

GET /v5/analytics

Allowed keys: AllowedKey::Country, AllowedKey::AdSlotType

GET /v5/analytics/for-publisher (auth required)

Returns all analytics where the currently authenticated address Auth.uid is a publisher.

All ALLOWED_KEYS are allowed for this route.

GET /v5/analytics/for-advertiser (auth required)

Returns all analytics where the currently authenticated address Auth.uid is an advertiser.

All ALLOWED_KEYS are allowed for this route.

GET /v5/analytics/for-admin (auth required)

Admin access to the analytics with no restrictions on the keys for filtering.

All ALLOWED_KEYS are allowed for admins.

Admin addresses are configured in the Config.admins

Modules

campaign

/v5/campaign routes

channel

/v5/channel routes

routers

This module contains all the Sentry REST API routers.

Functions

get_analytics

GET /v5/analytics request with query parameters: primitives::analytics::AnalyticsQuery.

get_cfg

GET /cfg request