Struct sentry::ClientOptions

pub struct ClientOptions {
    pub dsn: Option<Dsn>,
    pub transport: Box<TransportFactory>,
    pub in_app_include: Vec<&'static str>,
    pub in_app_exclude: Vec<&'static str>,
    pub extra_border_frames: Vec<&'static str>,
    pub max_breadcrumbs: usize,
    pub trim_backtraces: bool,
    pub release: Option<Cow<'static, str>>,
    pub repos: HashMap<String, RepoReference>,
    pub environment: Option<Cow<'static, str>>,
    pub server_name: Option<Cow<'static, str>>,
    pub sample_rate: f32,
    pub user_agent: Cow<'static, str>,
    pub http_proxy: Option<Cow<'static, str>>,
    pub https_proxy: Option<Cow<'static, str>>,
    pub shutdown_timeout: Duration,
    pub debug: bool,
    pub attach_stacktrace: bool,
    pub send_default_pii: bool,
    pub before_send: Option<Arc<Box<Fn(Event<'static>) -> Option<Event<'static>> + Send + Sync>>>,
    pub before_breadcrumb: Option<Arc<Box<Fn(Breadcrumb) -> Option<Breadcrumb> + Send + Sync>>>,
}

Configuration settings for the client.

Fields
dsn: Option<Dsn>

The DSN to use. If not set the client is effectively disabled.

transport: Box<TransportFactory>

The transport to use.

This is typically either a boxed function taking the client options by reference and returning a Transport, a boxed Arc<Transport> or alternatively the DefaultTransportFactory.

in_app_include: Vec<&'static str>

module prefixes that are always considered in_app

in_app_exclude: Vec<&'static str>

module prefixes that are never in_app

extra_border_frames: Vec<&'static str>

border frames which indicate a border from a backtrace to useless internals. Some are automatically included.

max_breadcrumbs: usize

Maximum number of breadcrumbs (0 to disable feature).

trim_backtraces: bool

Automatically trim backtraces of junk before sending.

release: Option<Cow<'static, str>>

The release to be sent with events.

repos: HashMap<String, RepoReference>

The repos to send along with the events.

environment: Option<Cow<'static, str>>

The environment to be sent with events.

server_name: Option<Cow<'static, str>>

The server name to be reported.

sample_rate: f32

The sample rate for event submission (0.0 - 1.0, defaults to 1.0)

user_agent: Cow<'static, str>

The user agent that should be reported.

http_proxy: Option<Cow<'static, str>>

An optional HTTP proxy to use.

This will default to the http_proxy environment variable.

https_proxy: Option<Cow<'static, str>>

An optional HTTPS proxy to use.

This will default to the HTTPS_PROXY environment variable or http_proxy if that one exists.

shutdown_timeout: Duration

The timeout on client drop for draining events on shutdown.

debug: bool

Enables debug mode.

attach_stacktrace: bool

Attaches stacktraces to messages.

send_default_pii: bool

If turned on some default PII informat is attached.

before_send: Option<Arc<Box<Fn(Event<'static>) -> Option<Event<'static>> + Send + Sync>>>

Before send callback.

before_breadcrumb: Option<Arc<Box<Fn(Breadcrumb) -> Option<Breadcrumb> + Send + Sync>>>

Before breadcrumb add callback.